+++ docs/source/changelog.rst
@@ -6,6 +6,9 @@ pycite's changelog
 
 
 * 
+  Added ``split_authors`` a simple method to clean and abbreviate author names. 
+
+* 
   Fixed issues with actions not running on GitHub.
 
 *