+++ docs/source/changelog.rst
@@ -6,6 +6,9 @@ pycite's changelog
 
 
 * 
+  Volumes no longer have the leading "v" attached. 
+
+* 
   Added ``split_authors`` a simple method to clean and abbreviate author names. 
 
 *